+++ src/main/java/roomescape/dto/AddReservationDto.java
@@ -3,7 +3,7 @@ package roomescape.dto;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public record AddReservationDto(String name, LocalDate date, LocalTime time){
+public record AddReservationDto(String name, LocalDate date, LocalTime time) {
 
 }
