+++ src/main/java/roomescape/exception/InvalidReservationRequest.java
@@ -2,7 +2,7 @@ package roomescape.exception;
 
 public class InvalidReservationRequest extends IllegalArgumentException {
 
-    public InvalidReservationRequest(){
+    public InvalidReservationRequest() {
         super("유효하지 않은 예약 요청입니다.");
     }
 }
