+++ src/main/resources/templates/admin/index.html
@@ -14,7 +14,8 @@
     <a class="navbar-brand" href="/">
         <img alt="LOGO" src="/image/admin-logo.png" style="height: 40px;">
     </a>
-    <button aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation" class="navbar-toggler"
+    <button aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation"
+            class="navbar-toggler"
             data-target="#navbarSupportedContent" data-toggle="collapse" type="button">
         <span class="navbar-toggler-icon"></span>
     </button>
